Separated ItemThumbnailUI with inheritance
