+++ Assets/Prefabs/UI/Shop Furniture Item.prefab.meta
