+++ Assets/Scripts/UI/Item Thumbnail UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1797c339749e5c04a8f1b63e6459941f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
