+++ Assets/Scripts/UI/Item Thumbnail UI/DecorateListItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6fa54399d3aa4a242a60ba04ada4e562
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
