+++ Assets/Scripts/UI/Item Thumbnail UI/ItemThumbnailUI.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+public class ItemThumbnailUI : MonoBehaviour
+{
+    [SerializeField] private Image image;
+
+    protected InventoryItem inventoryItem;
+    protected FurnitureSO furnitureSO;
+
+    public void SetItem(InventoryItem inventoryItem)
+    {
+        this.inventoryItem = inventoryItem;
+        furnitureSO = DataPersistenceManager.Instance.AllFurnitureSO.Find(x => x.id == inventoryItem.id);
+        image.sprite = furnitureSO.thumbnail;
+    }
+}
