+++ Assets/Scripts/UI/Item Thumbnail UI/ShopFurnitureItem.cs
@@ -1,0 +1,7 @@
+public class ShopFurnitureItem : ItemThumbnailUI
+{
+    public void SelectItem()
+    {
+        ShopUIManager.Instance.SelectItem((furnitureSO, inventoryItem));
+    }
+}
