+++ Assets/Scripts/UI/Item Thumbnail UI/ShopFurnitureItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f52e7d95df84bd48aa4ebb84cae5027
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
