+++ Assets/Scripts/UI/Item Thumbnail UI/ShopTabItem.cs
@@ -1,0 +1,7 @@
+public class ShopTabItem : ItemThumbnailUI
+{
+    public void SetTab()
+    {
+        ShopUIManager.Instance.SetTab(furnitureSO.id);
+    }
+}
