+++ Assets/Scripts/UI/Item Thumbnail UI/ShopTabItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec323e06488eb7d44969c3f60f306cd7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
